NAM2/requirements.txt
# ── 직접 의존성 ───────────────────────────────────────────────
# Python 3.12 기준으로 검증됨

# API 서버
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.5

# AI 오케스트레이션 (Azure OpenAI 커넥터 포함)
semantic-kernel==1.40.0

# Azure AI Search (정부지원사업 RAG 검색)
azure-search-documents==11.6.0b9

# HTTP 클라이언트
requests==2.32.3

# 환경 변수
python-dotenv==1.2.2

# ── 핵심 전이 의존성 (버전 고정) ────────────────────────────
openai==2.26.0
azure-identity==1.25.2
aiohttp==3.13.3
